role_based_cmc/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
sqlalchemy==2.0.36
psycopg2==2.9.10
asyncpg==0.30.0
alembic==1.14.0
bcrypt==4.2.1
pyjwt==2.10.1
